app_db_tests/requirements.txt
uwsgi 
aioredis
asgiref
async-timeout
attrs
Twisted
autobahn
Automat
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
Deprecated
Django
django-braces
django-debug-toolbar
django-embed-video
django-redisboard
djangorestframework
hiredis
hyperlink
idna
incremental
msgpack
packaging
Pillow
psycopg2
pyasn1
pyasn1-modules
pycparser
pymemcache
pyOpenSSL
pytz
redis
requests
service-identity
six
sqlparse
Twisted
txaio
typing_extensions
urllib3
wrapt
zope.interface
